refactor: 공지사항 model barrel index.ts 제거
