+++ packages/client/src/entities/notices/model/notice.ts
